Everything is now working and all of the old parser classes are now gone. I need to start a massive code clean up project now. There are a lot of comments that need to be added and test that need to be created. I need to reorganize a lot of stuff. At least everything is now working.
