+++ Dxflib/Entities/HatchBuffer.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Dxflib.Entities
+﻿namespace Dxflib.Entities
 {
     public class HatchBuffer : EntityBuffer
     {
