+++ Dxflib/Entities/LwPolyLineBuffer.cs
@@ -12,7 +12,6 @@
 using System.Collections.Generic;
 using Dxflib.Geometry;
 using Dxflib.IO;
-using Dxflib.Parser;
 
 namespace Dxflib.Entities
 {
