+++ Dxflib/IO/GroupCodesBase.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Dxflib.IO
+﻿namespace Dxflib.IO
 {
     /// <summary>
     ///
