+++ Dxflib/IO/TaggedDataList.cs
@@ -1,37 +1,42 @@
-﻿using System;
+﻿// Dxflib
+// TaggedDataList.cs
+// 
+// ============================================================
+// 
+// Created: 2018-08-21
+// Last Updated: 2018-08-21-8:54 PM
+// By: Adam Renaud
+// 
+// ============================================================
+
 using System.Collections.Generic;
-using System.Text;
 
 namespace Dxflib.IO
 {
+
     public class TaggedDataList
     {
-        private List<TaggedData> _list;
+        private readonly List<TaggedData> _list;
 
         /// <summary>
-        /// 
         /// </summary>
         /// <param name="stringList"></param>
         public TaggedDataList(string[] stringList)
         {
-            _list = new List<TaggedData>(stringList.Length/2);
-            for ( int i = 0; i < stringList.Length - 2; i +=2 )
-            {
+            _list = new List<TaggedData>(stringList.Length / 2);
+            for ( var i = 0; i < stringList.Length - 2; i += 2 )
                 _list.Add(new TaggedData(stringList[i], stringList[i + 1]));
-            }
-                
         }
 
         /// <summary>
-        /// Total Length of the list
+        ///     Total Length of the list
         /// </summary>
         public int Length => _list.Count;
 
         /// <summary>
-        /// 
         /// </summary>
         /// <param name="index"></param>
         /// <returns></returns>
-        public TaggedData GetPair(int index) => _list[index];
+        public TaggedData GetPair(int index) { return _list[index]; }
     }
 }
